ファイルをMVCに分け、ViewControllerからAPI通信部分を切り分けた。ViewController内でextensionを用いて処理を切り出し可読性を上げた。
