+++ iOSEngineerCodeCheck/View/Assets.xcassets/AppIcon.appiconset/Contents.json
